doc: use @node-core/remark-lint for linting
