+++ .github/workflows/test-macos.yml
@@ -22,7 +22,6 @@ on:
       - tools/doc/**
       - tools/eslint-rules/**
       - tools/eslint/**
-      - tools/lint-md/**
       - typings/**
       - vcbuild.bat
       - .**
@@ -52,7 +51,6 @@ on:
       - tools/doc/**
       - tools/eslint-rules/**
       - tools/eslint/**
-      - tools/lint-md/**
       - typings/**
       - vcbuild.bat
       - .**
