+++ doc/api/errors.md
@@ -3497,7 +3497,7 @@ changes:
   - version:
      - v11.4.0
      - v10.15.0
-    commit: 186035243fad247e3955f
+    commit: 186035243fad247e3955fa0c202987cae99e82db
     pr-url: https://github.com/nodejs-private/node-private/pull/143
     description: Max header size in `http_parser` was set to 8 KiB.
 -->
