+++ doc/type-map.json
@@ -1,9 +1,10 @@
+
 {
   "AbortController": "globals.html#class-abortcontroller",
   "AbortSignal": "globals.html#class-abortsignal",
   "Algorithm": "webcrypto.html#class-algorithm",
   "AlgorithmIdentifier": "webcrypto.html#class-algorithmidentifier",
-  "AsyncHook": "async_hooks.html#async_hookscreatehookoptions",
+  "AsyncHook": "async_hooks.html#async_hookscreatehookcallbacks",
   "AsyncLocalStorage": "async_context.html#class-asynclocalstorage",
   "AsyncResource": "async_hooks.html#class-asyncresource",
   "AesCbcParams": "webcrypto.html#class-aescbcparams",
@@ -45,6 +46,8 @@
   "EventListener": "events.html#event-listener",
   "EventTarget": "events.html#class-eventtarget",
   "File": "buffer.html#class-file",
+  "EncapsulatedBits": "webcrypto.html#class-encapsulatedbits",
+  "EncapsulatedKey": "webcrypto.html#class-encapsulatedkey",
   "FileHandle": "fs.html#class-filehandle",
   "Handle": "net.html#serverlistenhandle-backlog-callback",
   "Hash": "crypto.html#class-hash",
@@ -80,6 +83,7 @@
   "ReadableStreamDefaultController": "webstreams.html#class-readablestreamdefaultcontroller",
   "ReadableStreamDefaultReader": "webstreams.html#class-readablestreamdefaultreader",
   "ModuleRequest": "vm.html#type-modulerequest",
+  "SyncCPUProfileHandle": "v8.html#class-synccpuprofilehandle",
   "DatabaseSync": "sqlite.html#class-databasesync",
   "RecordableHistogram": "perf_hooks.html#class-recordablehistogram-extends-histogram",
   "RsaHashedKeyAlgorithm": "webcrypto.html#class-rsahashedkeyalgorithm",
@@ -93,6 +97,7 @@
   "Disposable": "https://tc39.es/proposal-explicit-resource-management/#sec-disposable-interface",
   "Session": "sqlite.html#class-session",
   "StatementSync": "sqlite.html#class-statementsync",
+  "SQLTagStore": "sqlite.html#class-sqltagstore",
   "Stream": "stream.html#stream",
   "SubtleCrypto": "webcrypto.html#class-subtlecrypto",
   "TestsStream": "test.html#class-testsstream",
@@ -123,5 +128,6 @@
   "zlib options": "zlib.html#class-options",
   "zstd options": "zlib.html#class-zstdoptions",
   "HTTP/2 Headers Object": "http2.html#headers-object",
+  "HTTP/2 Raw Headers": "http2.html#raw-headers",
   "HTTP/2 Settings Object": "http2.html#settings-object"
 }
